@@ -35,27 +35,26 @@
     "@fortawesome/pro-solid-svg-icons": "^6.1.2",
     "@fortawesome/pro-thin-svg-icons": "^6.1.2",
     "@fortawesome/react-fontawesome": "^0.2.0",
-    "@reduxjs/toolkit": "^1.7.2",
     "@sentry/react": "^7.7.0",
     "@sentry/tracing": "^7.7.0",
     "@sentry/types": "^7.7.0",
     "animate.css": "^4.1.1",
     "axios": "^0.26.1",
+    "cmdk": "^0.1.19",
     "immer": "^9.0.15",
     "js-cookie": "^3.0.1",
     "lottie-react": "^2.3.1",
     "luxon": "^2.3.1",
     "prop-types": ">=15",
-    "react": "17.0.1",
+    "react": "^18.2.0",
     "react-dnd": "^15.1.1",
     "react-dnd-html5-backend": "^15.1.2",
-    "react-dom": "17.0.1",
+    "react-dom": "^18.2.0",
     "react-hook-form": "^7.27.0",
     "react-idle-timer": "^4.6.4",
     "react-is": ">= 16.8.0",
-    "react-modal": "^3.14.4",
-    "react-query": "^3.34.16",
-    "react-redux": "^7.2.6",
+    "react-modal": "^3.15.1",
+    "react-query": "^3.39.2",
     "react-router-dom": "6",
     "react-toastify": "^9.0.1",
     "react-tooltip": "^4.2.21",
@@ -77,8 +76,8 @@
     "@types/js-cookie": "^3.0.1",
     "@types/luxon": "^2.3.0",
     "@types/node": "^17.0.23",
-    "@types/react": "~17.0.21",
-    "@types/react-dom": "^17.0.14",
+    "@types/react": "^18.0.18",
+    "@types/react-dom": "^18.0.6",
     "@types/react-modal": "^3.13.1",
     "@types/react-native": "~0.64.12",
     "@types/styled-components": "^5.1.24",
